% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seabird.R
\name{read_seabird}
\alias{read_seabird}
\title{Read in Seabird .cnv files that have common parameters}
\usage{
read_seabird(autoname = TRUE, recursive = FALSE)
}
\arguments{
\item{autoname}{Should the function attempt to extract the names of the variables from raw data or should they be left blank and specified by the user using colnames()? Defaults to TRUE}

\item{recursive}{Should the function look into sub-folders of the current folder?}
}
\value{
A single dataframe containing all the cast dale and a label column indicating the filename of the .cnv file
}
\description{
This function is meant to import raw cnv files into an R data frame.
}

